requirements.txt
streamlit>=1.29.0
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.21.0
requests>=2.25.0

